+++ lib/chef/knife/rackspace_server_delete.rb
@@ -1,6 +1,7 @@
 #
 # Author:: Adam Jacob (<adam@opscode.com>)
-# Copyright:: Copyright (c) 2009 Opscode, Inc.
+# Author:: Matt Ray (<matt@opscode.com>)
+# Copyright:: Copyright (c) 2009-2012 Opscode, Inc.
 # License:: Apache License, Version 2.0
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -18,6 +19,10 @@
 
 require 'chef/knife/rackspace_base'
 
+# These two are needed for the '--purge' deletion case
+require 'chef/node'
+require 'chef/api_client'
+
 class Chef
   class Knife
     class RackspaceServerDelete < Knife
@@ -26,31 +31,63 @@ class Chef
 
       banner "knife rackspace server delete SERVER_ID [SERVER_ID] (options)"
 
-      def run
-        @name_args.each do |instance_id|
-          
-          server = connection.servers.get(instance_id)
+      option :purge,
+        :short => "-P",
+        :long => "--purge",
+        :boolean => true,
+        :default => false,
+        :description => "Destroy corresponding node and client on the Chef Server, in addition to destroying the Rackspace node itself. Assumes node and client have the same name as the server (if not, add the '--node-name' option)."
 
-          msg("Instance ID", server.id.to_s)
-          msg("Host ID", server.host_id)
-          msg("Name", server.name)
-          msg("Flavor", server.flavor.name)
-          msg("Image", server.image.name)
-          msg("Public DNS Name", server.addresses["public"][0])
-          msg("Private IP Address", server.addresses["private"][0])
+      option :chef_node_name,
+        :short => "-N NAME",
+        :long => "--node-name NAME",
+        :description => "The name of the node and client to delete, if it differs from the server name. Only has meaning when used with the '--purge' option."
 
-          puts "\n"
-          confirm("Do you really want to delete this server")
-
-          server.destroy
-
-          ui.warn("Deleted server #{server.id} named #{server.name}")
+      # Extracted from Chef::Knife.delete_object, because it has a
+      # confirmation step built in... By specifying the '--purge'
+      # flag (and also explicitly confirming the server destruction!)
+      # the user is already making their intent known.  It is not
+      # necessary to make them confirm two more times.
+      def destroy_item(klass, name, type_name)
+        begin
+          object = klass.load(name)
+          object.destroy
+          ui.warn("Deleted #{type_name} #{name}")
+        rescue Net::HTTPServerException
+          ui.warn("Could not find a #{type_name} named #{name} to delete!")
         end
       end
 
-      def msg(label, value)
-        if value && !value.empty?
-          puts "#{ui.color(label, :cyan)}: #{value}"
+      def run
+        @name_args.each do |instance_id|
+          begin
+            server = connection.servers.get(instance_id)
+            msg_pair("Instance ID", server.id.to_s)
+            msg_pair("Host ID", server.host_id)
+            msg_pair("Name", server.name)
+            msg_pair("Flavor", server.flavor.name)
+            msg_pair("Image", server.image.name)
+            msg_pair("Public IP Address", public_ip(server))
+            msg_pair("Private IP Address", private_ip(server))
+
+            puts "\n"
+            confirm("Do you really want to delete this server")
+
+            server.destroy
+
+            ui.warn("Deleted server #{server.id}")
+
+            if config[:purge]
+              thing_to_delete = config[:chef_node_name] || instance_id
+              destroy_item(Chef::Node, thing_to_delete, "node")
+              destroy_item(Chef::ApiClient, thing_to_delete, "client")
+            else
+              ui.warn("Corresponding node and client for the #{instance_id} server were not deleted and remain registered with the Chef Server")
+            end
+
+          rescue NoMethodError
+            ui.error("Could not locate server '#{instance_id}'.")
+          end
         end
       end
     end